==> scripts/99-import_packages.R <==
#### Preamble ####
# Purpose: Installs packages needed to run files in "scripts" directory and "paper/paper.qmd"
# Author: Sima Shmuylovich
# Date: 15 February 2024
# Contact: sima.shmuylovich@mail.utoronto.ca
# License: MIT
# Pre-requisites: N/A

#### Workspace setup ####
install.packages("rmarkdown") 
install.packages("here") 
install.packages("rstanarm") 
install.packages("tidyverse") 
install.packages("janitor") 
install.packages("tidyr") 
install.packages("knitr")
install.packages("dplyr")
install.packages("ggplot2")
install.packages("kableExtra")
